repo rearrangement, renamings
