--- Sources/SwiftyModel/SwiftyModel.swift
@@ -1,6 +1,0 @@
-public struct SwiftyModel {
-    public private(set) var text = "Hello, World!"
-
-    public init() {
-    }
-}
