+++ Tests/SwiftyModelTests/SwiftyModelTests.swift
@@ -6,6 +6,6 @@ final class SwiftyModelTests: XCTestCase {
         // This is an example of a functional test case.
         // Use XCTAssert and related functions to verify your tests produce the correct
         // results.
-        XCTAssertEqual(SwiftyModel().text, "Hello, World!")
+        
     }
 }
